refactor: 관리자 등급 확인하는 부분 훅 분리
