+++ apps/admin/pages/seller/account.tsx
@@ -1,34 +1,16 @@
-import { useEffect } from 'react';
-import { Box, Heading, Text, useToast } from '@chakra-ui/react';
+import { Box, Heading, Text } from '@chakra-ui/react';
 import { AdminAccountList } from '@project-lc/components-admin/AdminAccountList';
 import { AdminPageLayout } from '@project-lc/components-admin/AdminPageLayout';
-import {
-  useAdminSettlementInfo,
-  useAdminPrivacyApproachHistoryMutation,
-  useProfile,
-} from '@project-lc/hooks';
-import { useRouter } from 'next/router';
+import { useAdminSettlementInfo, useCheckAdminClass } from '@project-lc/hooks';
 
 export function SellerAccountList(): JSX.Element {
   const { data: settlementData } = useAdminSettlementInfo();
-  const { mutateAsync } = useAdminPrivacyApproachHistoryMutation();
-  const { data: profile, isLoading } = useProfile();
-  const router = useRouter();
-  const toast = useToast();
 
-  if (!isLoading && !['super', 'full'].includes(profile.adminClass)) {
-    toast({
-      title: '권한없는 계정',
-      status: 'error',
-    });
-    router.push('/admin');
-  }
-
-  useEffect(() => {
-    if (!isLoading && ['super', 'full'].includes(profile.adminClass)) {
-      mutateAsync({ infoType: 'sellerSettlementAccount', actionType: 'view' });
-    }
-  }, []);
+  useCheckAdminClass({
+    adminClasses: ['super', 'full'],
+    infoType: 'sellerSettlementAccount',
+    actionType: 'view',
+  });
 
   return (
     <AdminPageLayout>
