+++ apps/admin/pages/seller/business-registration.tsx
@@ -1,29 +1,14 @@
-import { useEffect } from 'react';
-import { Box, Heading, Text, useToast } from '@chakra-ui/react';
+import { Box, Heading, Text } from '@chakra-ui/react';
 import { AdminBusinessRegistrationList } from '@project-lc/components-admin/AdminBusinessRegistrationList';
 import { AdminPageLayout } from '@project-lc/components-admin/AdminPageLayout';
-import { useAdminPrivacyApproachHistoryMutation, useProfile } from '@project-lc/hooks';
-import { useRouter } from 'next/router';
+import { useCheckAdminClass } from '@project-lc/hooks';
 
 export function SellerBusinessRegistration(): JSX.Element {
-  const { mutateAsync } = useAdminPrivacyApproachHistoryMutation();
-  const { data: profile, isLoading } = useProfile();
-  const router = useRouter();
-  const toast = useToast();
-
-  if (!isLoading && !['super', 'full'].includes(profile.adminClass)) {
-    toast({
-      title: '권한없는 계정',
-      status: 'error',
-    });
-    router.push('/admin');
-  }
-
-  useEffect(() => {
-    if (!isLoading && ['super', 'full'].includes(profile.adminClass)) {
-      mutateAsync({ infoType: 'sellerBusinessRegistration', actionType: 'view' });
-    }
-  }, []);
+  useCheckAdminClass({
+    adminClasses: ['super', 'full'],
+    infoType: 'sellerBusinessRegistration',
+    actionType: 'view',
+  });
 
   return (
     <AdminPageLayout>
